form logic to create session
